libs/hmpb: Factor out vote creation logic for fixtures
